refactor(data-point): Rename ReducerAsArray to ReducerParallel
