+++ packages/data-point/lib/reducer-types/reducer-helpers/reducer-parallel/index.js
